requirements.txt
Pillow==9.2.0
flask==3.0.3